# feat(scope-manager): only call Object.entries once for each lib