--- a/packages/scope-manager/src/lib/decorators.legacy.ts
+++ b/packages/scope-manager/src/lib/decorators.legacy.ts
@@ -3,13 +3,16 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const decorators_legacy = {
-  ClassDecorator: TYPE,
-  MethodDecorator: TYPE,
-  ParameterDecorator: TYPE,
-  PropertyDecorator: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const decorators_legacy: LibDefinition = {
+  libs: [],
+  variables: [
+    ['ClassDecorator', TYPE],
+    ['PropertyDecorator', TYPE],
+    ['MethodDecorator', TYPE],
+    ['ParameterDecorator', TYPE],
+  ],
+};