--- a/packages/scope-manager/src/lib/decorators.ts
+++ b/packages/scope-manager/src/lib/decorators.ts
@@ -3,21 +3,24 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const decorators = {
-  ClassAccessorDecoratorContext: TYPE,
-  ClassAccessorDecoratorResult: TYPE,
-  ClassAccessorDecoratorTarget: TYPE,
-  ClassDecoratorContext: TYPE,
-  ClassFieldDecoratorContext: TYPE,
-  ClassGetterDecoratorContext: TYPE,
-  ClassMemberDecoratorContext: TYPE,
-  ClassMethodDecoratorContext: TYPE,
-  ClassSetterDecoratorContext: TYPE,
-  DecoratorContext: TYPE,
-  DecoratorMetadata: TYPE,
-  DecoratorMetadataObject: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const decorators: LibDefinition = {
+  libs: [],
+  variables: [
+    ['ClassMemberDecoratorContext', TYPE],
+    ['DecoratorContext', TYPE],
+    ['DecoratorMetadataObject', TYPE],
+    ['DecoratorMetadata', TYPE],
+    ['ClassDecoratorContext', TYPE],
+    ['ClassMethodDecoratorContext', TYPE],
+    ['ClassGetterDecoratorContext', TYPE],
+    ['ClassSetterDecoratorContext', TYPE],
+    ['ClassAccessorDecoratorContext', TYPE],
+    ['ClassAccessorDecoratorTarget', TYPE],
+    ['ClassAccessorDecoratorResult', TYPE],
+    ['ClassFieldDecoratorContext', TYPE],
+  ],
+};