--- a/packages/scope-manager/src/lib/dom.asynciterable.ts
+++ b/packages/scope-manager/src/lib/dom.asynciterable.ts
@@ -3,13 +3,16 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const dom_asynciterable = {
-  FileSystemDirectoryHandle: TYPE,
-  FileSystemDirectoryHandleAsyncIterator: TYPE,
-  ReadableStream: TYPE,
-  ReadableStreamAsyncIterator: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const dom_asynciterable: LibDefinition = {
+  libs: [],
+  variables: [
+    ['FileSystemDirectoryHandleAsyncIterator', TYPE],
+    ['FileSystemDirectoryHandle', TYPE],
+    ['ReadableStreamAsyncIterator', TYPE],
+    ['ReadableStream', TYPE],
+  ],
+};