--- a/packages/scope-manager/src/lib/es2015.core.ts
+++ b/packages/scope-manager/src/lib/es2015.core.ts
@@ -3,30 +3,33 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const es2015_core = {
-  Array: TYPE,
-  ArrayConstructor: TYPE,
-  DateConstructor: TYPE,
-  Float32Array: TYPE,
-  Float64Array: TYPE,
-  Function: TYPE,
-  Int8Array: TYPE,
-  Int16Array: TYPE,
-  Int32Array: TYPE,
-  Math: TYPE,
-  NumberConstructor: TYPE,
-  ObjectConstructor: TYPE,
-  ReadonlyArray: TYPE,
-  RegExp: TYPE,
-  RegExpConstructor: TYPE,
-  String: TYPE,
-  StringConstructor: TYPE,
-  Uint8Array: TYPE,
-  Uint8ClampedArray: TYPE,
-  Uint16Array: TYPE,
-  Uint32Array: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2015_core: LibDefinition = {
+  libs: [],
+  variables: [
+    ['Array', TYPE],
+    ['ArrayConstructor', TYPE],
+    ['DateConstructor', TYPE],
+    ['Function', TYPE],
+    ['Math', TYPE],
+    ['NumberConstructor', TYPE],
+    ['ObjectConstructor', TYPE],
+    ['ReadonlyArray', TYPE],
+    ['RegExp', TYPE],
+    ['RegExpConstructor', TYPE],
+    ['String', TYPE],
+    ['StringConstructor', TYPE],
+    ['Int8Array', TYPE],
+    ['Uint8Array', TYPE],
+    ['Uint8ClampedArray', TYPE],
+    ['Int16Array', TYPE],
+    ['Uint16Array', TYPE],
+    ['Int32Array', TYPE],
+    ['Uint32Array', TYPE],
+    ['Float32Array', TYPE],
+    ['Float64Array', TYPE],
+  ],
+};