--- a/packages/scope-manager/src/lib/es2015.promise.ts
+++ b/packages/scope-manager/src/lib/es2015.promise.ts
@@ -3,10 +3,11 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const es2015_promise = {
-  PromiseConstructor: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2015_promise: LibDefinition = {
+  libs: [],
+  variables: [['PromiseConstructor', TYPE]],
+};