--- a/packages/scope-manager/src/lib/es2015.ts
+++ b/packages/scope-manager/src/lib/es2015.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { es5 } from './es5';
 import { es2015_collection } from './es2015.collection';
@@ -16,15 +16,18 @@ import { es2015_reflect } from './es2015.reflect';
 import { es2015_symbol } from './es2015.symbol';
 import { es2015_symbol_wellknown } from './es2015.symbol.wellknown';
 
-export const es2015 = {
-  ...es5,
-  ...es2015_core,
-  ...es2015_collection,
-  ...es2015_iterable,
-  ...es2015_generator,
-  ...es2015_promise,
-  ...es2015_proxy,
-  ...es2015_reflect,
-  ...es2015_symbol,
-  ...es2015_symbol_wellknown,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2015: LibDefinition = {
+  libs: [
+    es5,
+    es2015_core,
+    es2015_collection,
+    es2015_iterable,
+    es2015_generator,
+    es2015_promise,
+    es2015_proxy,
+    es2015_reflect,
+    es2015_symbol,
+    es2015_symbol_wellknown,
+  ],
+  variables: [],
+};