--- a/packages/scope-manager/src/lib/es2016.full.ts
+++ b/packages/scope-manager/src/lib/es2016.full.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { dom } from './dom';
 import { dom_iterable } from './dom.iterable';
@@ -11,10 +11,7 @@ import { es2016 } from './es2016';
 import { scripthost } from './scripthost';
 import { webworker_importscripts } from './webworker.importscripts';
 
-export const es2016_full = {
-  ...es2016,
-  ...dom,
-  ...webworker_importscripts,
-  ...scripthost,
-  ...dom_iterable,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2016_full: LibDefinition = {
+  libs: [es2016, dom, webworker_importscripts, scripthost, dom_iterable],
+  variables: [],
+};