--- a/packages/scope-manager/src/lib/es2017.full.ts
+++ b/packages/scope-manager/src/lib/es2017.full.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { dom } from './dom';
 import { dom_iterable } from './dom.iterable';
@@ -11,10 +11,7 @@ import { es2017 } from './es2017';
 import { scripthost } from './scripthost';
 import { webworker_importscripts } from './webworker.importscripts';
 
-export const es2017_full = {
-  ...es2017,
-  ...dom,
-  ...webworker_importscripts,
-  ...scripthost,
-  ...dom_iterable,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2017_full: LibDefinition = {
+  libs: [es2017, dom, webworker_importscripts, scripthost, dom_iterable],
+  variables: [],
+};