--- a/packages/scope-manager/src/lib/es2017.sharedmemory.ts
+++ b/packages/scope-manager/src/lib/es2017.sharedmemory.ts
@@ -3,17 +3,18 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE, TYPE_VALUE } from './base-config';
 import { es2015_symbol } from './es2015.symbol';
 import { es2015_symbol_wellknown } from './es2015.symbol.wellknown';
 
-export const es2017_sharedmemory = {
-  ...es2015_symbol,
-  ...es2015_symbol_wellknown,
-  ArrayBufferTypes: TYPE,
-  Atomics: TYPE_VALUE,
-  SharedArrayBuffer: TYPE_VALUE,
-  SharedArrayBufferConstructor: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2017_sharedmemory: LibDefinition = {
+  libs: [es2015_symbol, es2015_symbol_wellknown],
+  variables: [
+    ['SharedArrayBuffer', TYPE_VALUE],
+    ['SharedArrayBufferConstructor', TYPE],
+    ['ArrayBufferTypes', TYPE],
+    ['Atomics', TYPE_VALUE],
+  ],
+};