--- a/packages/scope-manager/src/lib/es2017.ts
+++ b/packages/scope-manager/src/lib/es2017.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { es2016 } from './es2016';
 import { es2017_arraybuffer } from './es2017.arraybuffer';
@@ -14,13 +14,16 @@ import { es2017_sharedmemory } from './es2017.sharedmemory';
 import { es2017_string } from './es2017.string';
 import { es2017_typedarrays } from './es2017.typedarrays';
 
-export const es2017 = {
-  ...es2016,
-  ...es2017_arraybuffer,
-  ...es2017_date,
-  ...es2017_intl,
-  ...es2017_object,
-  ...es2017_sharedmemory,
-  ...es2017_string,
-  ...es2017_typedarrays,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2017: LibDefinition = {
+  libs: [
+    es2016,
+    es2017_arraybuffer,
+    es2017_date,
+    es2017_intl,
+    es2017_object,
+    es2017_sharedmemory,
+    es2017_string,
+    es2017_typedarrays,
+  ],
+  variables: [],
+};