--- a/packages/scope-manager/src/lib/es2017.typedarrays.ts
+++ b/packages/scope-manager/src/lib/es2017.typedarrays.ts
@@ -3,18 +3,21 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const es2017_typedarrays = {
-  Float32ArrayConstructor: TYPE,
-  Float64ArrayConstructor: TYPE,
-  Int8ArrayConstructor: TYPE,
-  Int16ArrayConstructor: TYPE,
-  Int32ArrayConstructor: TYPE,
-  Uint8ArrayConstructor: TYPE,
-  Uint8ClampedArrayConstructor: TYPE,
-  Uint16ArrayConstructor: TYPE,
-  Uint32ArrayConstructor: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2017_typedarrays: LibDefinition = {
+  libs: [],
+  variables: [
+    ['Int8ArrayConstructor', TYPE],
+    ['Uint8ArrayConstructor', TYPE],
+    ['Uint8ClampedArrayConstructor', TYPE],
+    ['Int16ArrayConstructor', TYPE],
+    ['Uint16ArrayConstructor', TYPE],
+    ['Int32ArrayConstructor', TYPE],
+    ['Uint32ArrayConstructor', TYPE],
+    ['Float32ArrayConstructor', TYPE],
+    ['Float64ArrayConstructor', TYPE],
+  ],
+};