--- a/packages/scope-manager/src/lib/es2018.asyncgenerator.ts
+++ b/packages/scope-manager/src/lib/es2018.asyncgenerator.ts
@@ -3,14 +3,16 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 import { es2018_asynciterable } from './es2018.asynciterable';
 
-export const es2018_asyncgenerator = {
-  ...es2018_asynciterable,
-  AsyncGenerator: TYPE,
-  AsyncGeneratorFunction: TYPE,
-  AsyncGeneratorFunctionConstructor: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2018_asyncgenerator: LibDefinition = {
+  libs: [es2018_asynciterable],
+  variables: [
+    ['AsyncGenerator', TYPE],
+    ['AsyncGeneratorFunction', TYPE],
+    ['AsyncGeneratorFunctionConstructor', TYPE],
+  ],
+};