--- a/packages/scope-manager/src/lib/es2018.full.ts
+++ b/packages/scope-manager/src/lib/es2018.full.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { dom } from './dom';
 import { dom_asynciterable } from './dom.asynciterable';
@@ -12,11 +12,14 @@ import { es2018 } from './es2018';
 import { scripthost } from './scripthost';
 import { webworker_importscripts } from './webworker.importscripts';
 
-export const es2018_full = {
-  ...es2018,
-  ...dom,
-  ...webworker_importscripts,
-  ...scripthost,
-  ...dom_iterable,
-  ...dom_asynciterable,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2018_full: LibDefinition = {
+  libs: [
+    es2018,
+    dom,
+    webworker_importscripts,
+    scripthost,
+    dom_iterable,
+    dom_asynciterable,
+  ],
+  variables: [],
+};