--- a/packages/scope-manager/src/lib/es2018.intl.ts
+++ b/packages/scope-manager/src/lib/es2018.intl.ts
@@ -3,10 +3,11 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE_VALUE } from './base-config';
 
-export const es2018_intl = {
-  Intl: TYPE_VALUE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2018_intl: LibDefinition = {
+  libs: [],
+  variables: [['Intl', TYPE_VALUE]],
+};