--- a/packages/scope-manager/src/lib/es2019.array.ts
+++ b/packages/scope-manager/src/lib/es2019.array.ts
@@ -3,12 +3,15 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const es2019_array = {
-  Array: TYPE,
-  FlatArray: TYPE,
-  ReadonlyArray: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2019_array: LibDefinition = {
+  libs: [],
+  variables: [
+    ['FlatArray', TYPE],
+    ['ReadonlyArray', TYPE],
+    ['Array', TYPE],
+  ],
+};