--- a/packages/scope-manager/src/lib/es2019.object.ts
+++ b/packages/scope-manager/src/lib/es2019.object.ts
@@ -3,12 +3,12 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 import { es2015_iterable } from './es2015.iterable';
 
-export const es2019_object = {
-  ...es2015_iterable,
-  ObjectConstructor: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2019_object: LibDefinition = {
+  libs: [es2015_iterable],
+  variables: [['ObjectConstructor', TYPE]],
+};