--- a/packages/scope-manager/src/lib/es2019.symbol.ts
+++ b/packages/scope-manager/src/lib/es2019.symbol.ts
@@ -3,10 +3,11 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { TYPE } from './base-config';
 
-export const es2019_symbol = {
-  Symbol: TYPE,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2019_symbol: LibDefinition = {
+  libs: [],
+  variables: [['Symbol', TYPE]],
+};