--- a/packages/scope-manager/src/lib/es2019.ts
+++ b/packages/scope-manager/src/lib/es2019.ts
@@ -3,7 +3,7 @@
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
 // npx nx generate-lib repo
 
-import type { ImplicitLibVariableOptions } from '../variable';
+import type { LibDefinition } from '../variable';
 
 import { es2018 } from './es2018';
 import { es2019_array } from './es2019.array';
@@ -12,11 +12,14 @@ import { es2019_object } from './es2019.object';
 import { es2019_string } from './es2019.string';
 import { es2019_symbol } from './es2019.symbol';
 
-export const es2019 = {
-  ...es2018,
-  ...es2019_array,
-  ...es2019_object,
-  ...es2019_string,
-  ...es2019_symbol,
-  ...es2019_intl,
-} as Record<string, ImplicitLibVariableOptions>;
+export const es2019: LibDefinition = {
+  libs: [
+    es2018,
+    es2019_array,
+    es2019_object,
+    es2019_string,
+    es2019_symbol,
+    es2019_intl,
+  ],
+  variables: [],
+};