… packages/scope-manager/src/lib/dom.iterable.ts
… packages/scope-manager/src/lib/dom.ts
… packages/scope-manager/src/lib/es2015.iterable.ts
… packages/scope-manager/src/lib/es2015.symbol.wellknown.ts
… packages/scope-manager/src/lib/es2020.bigint.ts
… packages/scope-manager/src/lib/es2020.date.ts
… packages/scope-manager/src/lib/es2020.full.ts
… packages/scope-manager/src/lib/es2020.intl.ts
… packages/scope-manager/src/lib/es2020.number.ts
… packages/scope-manager/src/lib/es2020.promise.ts
… packages/scope-manager/src/lib/es2020.sharedmemory.ts
… packages/scope-manager/src/lib/es2020.string.ts
… packages/scope-manager/src/lib/es2020.symbol.wellknown.ts
… packages/scope-manager/src/lib/es2020.ts
… packages/scope-manager/src/lib/es2021.full.ts
… packages/scope-manager/src/lib/es2021.intl.ts
… packages/scope-manager/src/lib/es2021.promise.ts
… packages/scope-manager/src/lib/es2021.string.ts
… packages/scope-manager/src/lib/es2021.ts
… packages/scope-manager/src/lib/es2021.weakref.ts
… packages/scope-manager/src/lib/es2022.array.ts
… packages/scope-manager/src/lib/es2022.error.ts
… packages/scope-manager/src/lib/es2022.full.ts
… packages/scope-manager/src/lib/es2022.intl.ts
… packages/scope-manager/src/lib/es2022.object.ts
… packages/scope-manager/src/lib/es2022.regexp.ts
… packages/scope-manager/src/lib/es2022.string.ts
… packages/scope-manager/src/lib/es2022.ts
… packages/scope-manager/src/lib/es2023.array.ts
… packages/scope-manager/src/lib/es2023.collection.ts