@dmichon-msft 2025-01-14
Since you gave me license to restructure the objects from the individual `lib` files, I went ahead and separated out the concept of "references to other libs" from individual variable definitions, so that in the (unfortunately rather common) case that developers have specified their `lib` array in such a way that a given lib is included multiple times, we only iterate the variables from it once (per source file, still).