@bradzacher 2025-02-25
I'm happy to go with a bigger refactor of how all of this works.
We should be able to mitigate the cost of loading libs and the cost of defining variables using this approach which is nice.

Just need to get the tests passing and we can move forward with this!