@JoshuaKGoldberg 2025-03-03
Before this change, running https://github.com/typescript-eslint/performance:

```plaintext
┌───────┬───────────────────────┬───────────────────────┐
│ files │ project (even layout) │ service (even layout) │
├───────┼───────────────────────┼───────────────────────┤
│ 1024  │ '2.760 s ±  0.006 s'  │ '3.125 s ±  0.011 s'  │
└───────┴───────────────────────┴───────────────────────┘
```

After:

```plaintext
┌───────┬───────────────────────┬───────────────────────┐
│ files │ project (even layout) │ service (even layout) │
├───────┼───────────────────────┼───────────────────────┤
│ 1024  │ '2.478 s ±  0.022 s'  │ '2.829 s ±  0.020 s'  │
└───────┴───────────────────────┴───────────────────────┘
```